requirements.txt
flask
flask-cors
ultralytics
pillow
torch
torchvision
gunicorn
